move files into respective folders
