+++ frontend/components/Connection.js
