+++ frontend/screens/AppTabs.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs'; 
 import ProfileScreen from './ProfileScreen';
 import EmptyScreen from './EmptyScreen';
-import ConnectionsScreen from '../connections/ConnectionsScreen';
+import ConnectionsScreen from './ConnectionsScreen';
 import {HomeOutline, FriendsOutline, PlusOutline, BellOutline, MailOutline} from '../icons/index';
 
 const Tab = createBottomTabNavigator();
